+++ Packages/Teem/Dataflow/Modules/Unu/UnuCCadj.cc
@@ -110,7 +110,7 @@ UnuCCadj::execute()
   out->copy_properties(nrrd_handle.get_rep());
 
   // Copy the axis kinds
-  for (int i=0; i<nin->dim && i<nout->dim; i++) {
+  for (size_t i=0; i<nin->dim && i<nout->dim; i++) {
     nout->axis[i].kind = nin->axis[i].kind;
   }
